build: update Rust nightly to 2026-09-01 (#418)
* build: update Rust nightly and edition

* build: update platform dependencies for latest nightly

* build: minimize vendored nightly compatibility patches

* build: use upstream trapframe fix

* refactor: fold kernel synchronization into HAL

* ci: restore nightly compatibility

* build: refresh macro dependencies

* ci: install QEMU build dependencies

* ci: use packaged QEMU

* ci: avoid duplicate pull request runs

* build: update trapframe and enable FSGSBASE

* ci: remove scheduled builds and tests

* ci: make full checks manually dispatchable

* ci: cancel superseded pull request runs

* ci: replace deprecated GitHub Actions

* fix: initialize vfork child address space

* fix: handle all paths without panicking

+++ kernel-hal/src/bare/arch/aarch64/timer.rs
@@ -8,7 +8,7 @@ use tock_registers::interfaces::{Readable, Writeable};
 pub fn timer_now() -> Duration {
     unsafe { barrier::isb(barrier::SY) }
     let cur_cnt = CNTPCT_EL0.get() * 1_000_000_000;
-    let freq = CNTFRQ_EL0.get() as u64;
+    let freq = CNTFRQ_EL0.get();
     Duration::from_nanos(cur_cnt / freq)
 }
