+++ zircon-object/src/dev/pci/caps.rs
@@ -4,7 +4,7 @@ use crate::{ZxError, ZxResult};
 use alloc::boxed::Box;
 use core::convert::TryFrom;
 use kernel_hal::interrupt;
-use lock::Mutex;
+use kernel_hal::sync::Mutex;
 
 /// Enumeration for PCI capabilities.
 #[derive(Debug)]
@@ -149,7 +149,7 @@ impl PciCapPcie {
         let device_caps = cfg.read32_(base as usize + 0x4);
         PciCapPcie {
             version: caps & 0xF,
-            dev_type: PcieDeviceType::try_from(((caps >> 4) & 0xF) as u8).unwrap(),
+            dev_type: PcieDeviceType::try_from((caps >> 4) & 0xF).unwrap(),
             has_flr: ((device_caps >> 28) & 0x1) != 0,
         }
     }
